Issue 194 (#195)
* build pninexus by using conan with VS2019 on windows 11

* chage HDF5 lib version
* fixed min/max macro confliction
* changed 'and' to '&&' for backward compatibility

* issue _194

added PNINEXUS_BUILD_DOC option (default value on) in cmake file

* added windows job in workflow

* added msvc to path

* issue_194

install conan

* issue 194

build h5cpp first

* issue_194

fix syntax error

* issue_194

add h5cpp source

* build h5cpp

* test step build h5cpp

* test step build h5cpp

* test step build h5cpp

* test step build h5cpp

* test step build h5cpp

* updated conan file and added build pninexus

* fix syntax error in yaml

* change cmake version

* test different cmake version

* test windows build

* changed conanfile

* changed conanfile

* fix path error

* test build pninexus

* fix path error

* test build pninexus

* test build pninexus

* test build pninexus

* reenable linux jobs

* enable tests on windows

* fix error when run tests

* fix job run tests

* update conanfile to use shared HDF5
* copy h5cpp.dll to test folder

* fix syntax error in test.yml

* clean warnings

* disable warnings 4251 on windows
* reenabled linux jobs

+++ src/pni/nexus/xml/dataset_creation_list_builder.cpp
@@ -74,7 +74,7 @@ void DatasetCreationListBuilder::set_compression(hdf5::property::DatasetCreation
       use_shuffle = node.attribute("shuffle").data<hdf5::datatype::EBool>();
 
     if(node.has_attribute("rate"))
-      compression_rate = node.attribute("rate").data<size_t>();
+      compression_rate = static_cast<long>(node.attribute("rate").data<size_t>());
 
     if(use_compression)
     {
